+++ NEWS.md
@@ -5,6 +5,7 @@
 # Version 0.1.3
 - Keep num orden from IFN data
 - Data set Poblet tree data added
+- New function modifySoilParams()
 
 # Version 0.1.2
 - SoilGrids REST API relocated to https://rest.isric.org/soilgrids/v2.0/docs